Change UI organization, change approach to mining
The amount of resources in miners are not generated based upon a generation resource ledger, and are now instead set in their productivity.

Change up market information tooltip so that supply, demand, and SD ratio are showed in one unified table.

Remove supply from resource ledger because we take that from the market resource ledger anyway.