Some modifications to the build system.
Add option to add your own OpenAL

This is so that you don't have to delete CMakeCache.txt in the build folder every time the OpenAL project disappears.

Scripts were also written to set the icon, and the .o file was also added.
Add tests for engine

Put core tests in a separate folder, and add engine tests to it's own folder.

Change up how textures are freed so that it doesn't crash the test.

Add default constructor to Application, so that we can load it for a short while.