Refactor universe and script engine into a class (#92)
Unity the universe and script engine, that was previously defined in the application, to a new class, the Game class. This allows us to add more attributes and variables to the game without needing to add more arguments to classes and functions.

Make all systems retain the game object class, so that we don't need to pass the game argument over and over again.

Make the script runner become a system, so we don't have a separate class and system dealing with the scripts.