Upgrade the repo to use entt 3.10.0.
ResourceLedger was changed from a struct to a class.

Added a `LedgerMap` typedef.

The access specifier for std::map<entt::entity, double> was changed from public to private.

This avoids errors with entt emplacing a ResourceLedger or its subclasses into the registry.

ResourceDistribution was changed from making std::map<entt::entity, double> it's superclass to its subclass so it avoids the same issues.

ResourceDistribution was not changed in a similar manner to ResourceLedger because ResourceDistribution was less widely used.

Removed "Star System Viewer Debug" window because the API for ctx changed.